MyRatingService/requirements.txt
bidict==0.21.2
click==8.0.1
Flask==2.0.1
Flask-SocketIO==5.0.3
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector==2.2.9
python-engineio==4.2.0
python-socketio==5.3.0
Werkzeug==2.0.1
